sfs
sds

+++ App/Controllers/task/dashboard.php
@@ -14,9 +14,8 @@ require_once "../App/Models/user.php";
             $tasks = $taskModel->getAllTasksByUser($loggedInUser['id']);
 
         require "../App/Views/task/task-dashboard-view.php";
-        var_dump($_POST);
+        if(isset($_POST["id"])){
         $taskModel->delete($_POST["id"]);
-        header("Location: /task/dashboard");
-
-
+        header("Location: /task/calendar");
+        };
 ?>
